% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mpcr}
\alias{mpcr}
\title{Estimation of treatment effects in matched-pair cluster randomized
trials by calibrating covarite imbalance between clusters.}
\usage{
mpcr(datinput, arm = "tx", cluster = "team", pair = "sitenew",
  outcome = "sf36pcs32", X_nm_all = c("race", "ageatint", "hcc",
  "livesalone", "education", "s10", "sf36mcs", "sf36pcs", "gender", "h1"),
  X_nm_binary = c("livesalone", "education", "gender"), X_nm_cat = c("race",
  "s10", "h1"), X_nm_cont = c("ageatint", "hcc", "sf36mcs", "sf36pcs"))
}
\arguments{
\item{datinput}{The data set should have columns corresponding to the
primary outcome, treatment assignment, pair IDs, cluster IDs, covariates
used for covariate-calibration;}

\item{arm}{The name of treatment assignment indicator. For two-arm trials,
this variable takes value in {0,1}: 0 for control, 1 for treatment;}

\item{cluster}{The variable name for cluster IDs;}

\item{pair}{The variable name for pair IDs;}

\item{outcome}{The variable name for the primary outcome;}

\item{X_nm_all}{The vector of covariate names that enter the
covariate-calibrated analysis}

\item{X_nm_binary}{The vector of binary covariate names;}

\item{X_nm_cat}{The vector of categorical (>2 categories) covariate names;}

\item{X_nm_cont}{The vector of continuous covariate names.}
}
\value{
\itemize{
\item Tables:
\itemize{
        \item  Table 1 - cluster sample sizes;
          calibrated and uncalibrated outcome comparisons;
         \item Table 2 - check covariate imbalances within each pair;
         \item Table 3 -
             \itemize{
                  \item 1st level analysis: maximum likelihood estimate (MLE);
                   permutation tests;
                  \item 1st and 2nd level analysis:\cr
                      MLE; profile MLE;\cr
                      Bayes estimate with uniform shrinkage prior [link to paper];\cr
                      permutation tests.
             }
}

\item Figures:\cr

  Check second level dependence for crude analysis \cr
                     \deqn{\sqrt{v^{crude}_p} vs \delta^{crude}_p,}
                     and covariate-calibrated analysis
                     \deqn{\sqrt{v^{calibr}_p} vs \delta^{calibr}_p.}

}
}
\description{
Estimation of treatment effects in matched-pair cluster randomized
trials by calibrating covarite imbalance between clusters.
}
\references{
Wu, Z., Frangakis, C. E., Louis, T. A. and Scharfstein, D. O. (2014), Estimation of treatment effects in matched-pair cluster randomized trials by calibrating covariate imbalance between clusters. Biometrics. doi: 10.1111/biom.12214
}

