\name{MPR-package}
\alias{MPR-package}
\alias{MPR}
\docType{package}
\title{
  Multi Precision For R
}
\description{
  MPR offers 3 Different Precision For R (16-bit/32-bit/64-bit)
}
\details{
    MPR offers 3 Different Precision For R (16-bit/32-bit/64-bit)
    supporting Vectors and Matrix operation for Basic and Mathematical operations,and offering MultiPrecision operations.
}
\author{
Your David, email optional.

Maintainer: David Helmy <david.helmy@brightskiesinc.com>
}

\keyword{ package }
\seealso{
  README.md
}
\examples{

}
