\name{MPR Module Examples}
\alias{DataType}
\alias{Rcpp DataType Class}

\title{
  Rcpp Module Object for exporting C++ MPR Class to R as S4 object
}
\description{
  These function and objects are accessible from R via the Rcpp Modules mechanism
  which creates them based on the declaration in the C++ file.  
}
\seealso{
  The Rcpp Modules vignette.
}